# perf: use improved comparison algorithm in `no-identical-tests` rule