Spent some time this evening profiling eslint on the eslint repo. Noticed that this rule seems to take a little longer than I would've guessed and spend a bit of time with @jviide improving the equality algorithm used for this rule. The result is that the changes in this PR make the rule ~10x faster.

| Before | After |
|--:|--:|
| 453ms | **46ms** |